ai-service-python/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-community
langchain-ollama
chromadb
python-dotenv
pydantic>=2.0
pypdf
langchain-text-splitters